mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "mcp>=1.24.0",
    "python-dotenv>=1.2.1",
]
